Fire Rate added
The way I added it makes it where only 3 lasers can be present at once
at a certain rate.  So If I hold down the space bar it will fire at a
constant rate, however, It can be easily adapted to shoot more or switch
it to where they can shoot 3 as fast as they want to but only 3 can be
present.  its up to you.